+++ lists-ktx/src/main/kotlin/mathtools/lists/DoubleList.kt
