+++ lists-ktx/src/main/kotlin/mathtools/lists/LongList.kt
