--- lists/src/main/java/mathtools/lists/objects/package-info.java
@@ -1,5 +1,0 @@
-/**
- * General List operations on Object types
- * @author DK96-OS
- */
-package mathtools.lists.objects;
